students/app/requirements.txt
fastapi
sqlalchemy
python-jose
passlib==1.7.4
bcrypt==4.0.1
psycopg2-binary
PyJWT==2.10.1
aiokafka
kafka-python
confluent-kafka